I made some code for the grabbers
I this is just the Arduino to grabber communication. I don't have any of the input to the Arduino. Instead I have some booleans which should be set to true when the button is pressed and false whenever it isn't except for two buttons which I specify with comments in the code which should be toggles so if it is pressed it stays true even after released and is only false when the other of the two buttons is pressed. (this is to select which motor to spin when spinning).